Added tasks endpoints & images service w/ Cloudinary

+++ EmployeeAdministration/EmployeeAdministration.Domain/Enums/Roles.cs
@@ -1,6 +1,6 @@
 ﻿namespace EmployeeAdministration.Domain.Enums;
 
-public enum Roles : ushort
+public enum Roles : sbyte
 {
     Administrator = 1,
     Employee = 2
